+++ src/canisters/frontend/ic-asset/src/plumbing.rs
@@ -279,13 +279,14 @@ async fn upload_content_chunks(
         .map(|(i, data_chunk)| {
             create_chunk(canister, batch_id, data_chunk, semaphores).map_ok(move |chunk_id| {
                 println!(
-                    "  {}{} {}/{} ({} bytes) sha {}",
+                    "  {}{} {}/{} ({} bytes) sha {}{}",
                     &asset_descriptor.key,
                     content_encoding_descriptive_suffix(content_encoding),
                     i + 1,
                     count,
                     data_chunk.len(),
-                    hex::encode(sha256)
+                    hex::encode(sha256),
+                    &asset_descriptor.config
                 );
                 chunk_id
             })